# perf(core): replace `neo-async` with a minimal helper