… packages/rspack/package.json
… pnpm-lock.yaml